Restructure tests in reducer file
