+++ .gitignore
@@ -1,4 +1,5 @@
 node_modules/**/*
+.DS_Store
 .expo/*
 npm-debug.*
 *.jks
